extract likesPost service
