+++ App/Services/firebase/index.js
@@ -1,5 +1,7 @@
 // @flow
-import type {TUser} from '../../types/';
+import type {TUser, TLike, TChannel} from '../../types/';
+import {isSuccess, firebaseServiceResponse, handleServerError} from '../index';
+import {likesRef, channelsRef, likesRefUpdate, channelsRefUpdate} from './ref';
 
 export * from './auth';
 export * from './authErrorToMessage';
@@ -20,4 +22,23 @@ export const convertUserFromFirebaseToStore = (firebaseUser: TUser) => {
     refreshToken: firebaseUser.refreshToken,
     uid: firebaseUser.uid,
   };
+};
+
+export const likesPostToFirebase = async (channelId: string, uid: string, count: number) => {
+  const likeExist = await firebaseServiceResponse(likesRef.child(uid).orderByChild('channelId').equalTo(channelId).once('value'));
+  if (isSuccess(likeExist)) {
+    const like: TLike = likeExist.snapshot.val();
+    const likeKey: string = likeExist.snapshot.key;
+    handleServerError(likesRefUpdate(channelId, {[`/${likeKey}/count`]: like.count + count}));
+  } else {
+    handleServerError(likesRef.push({
+      channelId,
+      rank: 0,
+      count,
+    }));
+  }
+
+  const channelResponse = await firebaseServiceResponse(channelsRef.orderByChild('id').equalTo(channelId).once('value'));
+  const channel: TChannel = channelResponse.snapshot.val();
+  handleServerError(channelsRefUpdate(channelId, {[`/${channelId}/likeCount`]: channel.likeCount + count}));
 };
